+++ react/FetchPrevious.js
@@ -16,14 +16,21 @@ const FetchPrevious = () => {
   )
 
   const fetchMore = path(['fetchMore'], searchQuery)
+  const queryData = {
+    query: path(['variables', 'query'], searchQuery),
+    map: path(['variables', 'map'], searchQuery),
+    orderBy: path(['variables', 'orderBy'], searchQuery),
+    priceRange: path(['variables', 'priceRange'], searchQuery),
+  }
 
-  const { handleFetchMorePrevious, loading, from } = useFetchMore(
+  const { handleFetchMorePrevious, loading, from } = useFetchMore({
     page,
     recordsFiltered,
     maxItemsPerPage,
     fetchMore,
-    products
-  )
+    products,
+    queryData,
+  })
 
   return (
     <div