+++ react/hooks/useFetchMore.js
@@ -1,4 +1,4 @@
-import { useState, useRef, useCallback } from 'react'
+import { useState, useRef, useCallback, useEffect } from 'react'
 import { min, max } from 'ramda'
 import { useRuntime } from 'vtex.render-runtime'
 import {
@@ -102,13 +102,15 @@ const useFetchingMore = () => {
   return [stateValue, setFetchMore]
 }
 
-export const useFetchMore = (
-  page,
-  recordsFiltered,
-  maxItemsPerPage,
-  fetchMore,
-  products
-) => {
+export const useFetchMore = props => {
+  const {
+    page,
+    recordsFiltered,
+    maxItemsPerPage,
+    fetchMore,
+    products,
+    queryData: { query, map, orderBy, priceRange },
+  } = props
   const { setQuery } = useRuntime()
   const [currentPage, setCurrentPage] = useState(page)
   const [nextPage, setNextPage] = useState(page + 1)
@@ -116,12 +118,24 @@ export const useFetchMore = (
   const [currentFrom, setCurrentFrom] = useState((page - 1) * maxItemsPerPage)
   const [currentTo, setCurrentTo] = useState(currentFrom + maxItemsPerPage - 1)
   const [loading, setLoading] = useFetchingMore()
+  const isFirstRender = useRef(true)
   const fetchMoreLocked = useRef(false) // prevents the user from sending two requests at once
   /* this is a temporary solution to deal with unexpected 
   errors when the search result uses infinite scroll. 
   This should be removed once infinite scrolling is removed */
   const [infiniteScrollError, setInfiniteScrollError] = useState(false)
   const updateQueryError = useRef(false) //TODO: refactor this ref
+
+  useEffect(() => {
+    if (!isFirstRender.current) {
+      setCurrentPage(1)
+      setNextPage(2)
+      setPreviousPage(0)
+      setCurrentFrom(0)
+      setCurrentTo(maxItemsPerPage - 1)
+    }
+    isFirstRender.current = false
+  }, [maxItemsPerPage, query, map, orderBy, priceRange])
 
   const handleFetchMoreNext = async () => {
     const from = currentTo + 1